Fixed various issues
- added new tables for PostgreSQL, MS SQL and SQLite
- fixed an issue for adding rights
- removed HTML output during tests
- don't delete setup/update files if debug is enabled

+++ phpmyfaq/src/phpMyFAQ/Installer.php
@@ -915,8 +915,6 @@ class Installer
 
         $this->_system->setDatabase($db);
 
-        echo '<p>';
-
         // Erase any table before starting creating the required ones
         if (!System::isSqlite($dbSetup['dbType'])) {
             $this->_system->dropTables($uninst);
@@ -1010,8 +1008,6 @@ class Installer
             $faqInstanceElasticsearch = new Elasticsearch($configuration);
             $faqInstanceElasticsearch->createIndex();
         }
-
-        echo '</p>';
     }
 
     /**
@@ -1021,17 +1017,20 @@ class Installer
      */
     public function cleanUpFiles()
     {
-        // Remove 'index.php' file
-        if (@unlink(dirname($_SERVER['PATH_TRANSLATED']).'/index.php')) {
-            echo "<p class=\"alert alert-success\">The file <em>./setup/index.php</em> was deleted automatically.</p>\n";
-        } else {
-            echo "<p class=\"alert alert-danger\">Please delete the file <em>./setup/index.php</em> manually.</p>\n";
+        if (!DEBUG) {
+            // Remove 'index.php' file
+            if (@unlink(dirname($_SERVER['PATH_TRANSLATED']).'/index.php')) {
+                echo "<p class=\"alert alert-success\">The file <em>./setup/index.php</em> was deleted automatically.</p>\n";
+            } else {
+                echo "<p class=\"alert alert-danger\">Please delete the file <em>./setup/index.php</em> manually.</p>\n";
+            }
+            // Remove 'update.php' file
+            if (@unlink(dirname($_SERVER['PATH_TRANSLATED']).'/update.php')) {
+                echo "<p class=\"alert alert-success\">The file <em>./setup/update.php</em> was deleted automatically.</p>\n";
+            } else {
+                echo "<p class=\"alert alert-danger\">Please delete the file <em>./setup/update.php</em> manually.</p>\n";
+            }
         }
-        // Remove 'update.php' file
-        if (@unlink(dirname($_SERVER['PATH_TRANSLATED']).'/update.php')) {
-            echo "<p class=\"alert alert-success\">The file <em>./setup/update.php</em> was deleted automatically.</p>\n";
-        } else {
-            echo "<p class=\"alert alert-danger\">Please delete the file <em>./setup/update.php</em> manually.</p>\n";
-        }
+
     }
 }
